requirements.txt
pandas==2.2.2
numpy==2.0.1
scikit-learn==1.5.2
openai==1.59.7
python-dotenv==1.0.1
pydantic==2.7.4
markdown-pdf==1.3
anthropic==0.42.0
google-auth==2.37.0
requests==2.32.3
boto3==1.36.0
mistralai==1.5.0